requirements.txt
flask==3.0.3
openai==1.35.0
PyMuPDF==1.24.5
pytesseract==0.3.13
Pillow==10.4.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.134.0
python-dotenv==1.0.1
